requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary  # PostgreSQL driver
pydantic
pydantic-settings          # For BaseSettings
python-jose[cryptography]  # JWT
passlib[argon2]>=1.7.4     # Password hashing with Argon2 support
argon2-cffi>=21.3.0       # Explicit argon2 dependency
python-multipart           # File uploads (CSV)
pandas                     # CSV parsing
weasyprint                 # HTML-to-PDF generation
reportlab                  # PDF generation
jinja2                     # Template rendering
gunicorn                   # WSGI server for production
email-validator            # Email validation for pydantic
supabase>=2.3.0           # Supabase client for storage
boto3>=1.28.0             # AWS S3 client for Supabase storage
a2wsgi>=1.8.0             # ASGI to WSGI adapter for PythonAnywhere
